refactor: rename shared-module to shared
