+++ src/app/modules/about/about.module.ts
@@ -3,7 +3,7 @@ import { CommonModule } from '@angular/common';
 import { AboutMainComponent } from './components/about-main/about-main.component';
 import { AboutComponent } from './about.component';
 import { AboutRoutingModule } from './about.routing';
-import { SharedModule } from '@shared/shared-module.module';
+import { SharedModule } from '@shared/shared.module';
 import { AboutService } from './services/about.service';
