+++ src/app/modules/calendar/calendar.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
-import { SharedModule } from '@shared/shared-module.module';
+import { SharedModule } from '@shared/shared.module';
 
 import { CalendarRoutingModule } from './calendar.routing';
 import { DateService } from './services';
