+++ src/app/modules/contact/components/contact/contact.component.spec.ts
@@ -1,7 +1,7 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { ContactMainComponent } from './contact.component';
-import { SharedModule } from '@shared/shared-module.module';
+import { SharedModule } from '@shared/shared.module';
 
 describe('ContactComponent', () => {
   let component: ContactMainComponent;
