+++ src/app/modules/contact/contact.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { SharedModule } from '@shared/shared-module.module';
+import { SharedModule } from '@shared/shared.module';
 import { ContactMainComponent } from './components/contact/contact.component';
 import { ContactComponent } from './contact.component';
 import { ContactRoutingModule } from './contact.routing';
