+++ src/app/modules/core/services/dialog.service.spec.ts
@@ -1,7 +1,7 @@
 import { TestBed } from '@angular/core/testing';
 
 import { DialogService } from './dialog.service';
-import { SharedModule } from '@shared/shared-module.module';
+import { SharedModule } from '@shared/shared.module';
 
 describe('DialogService', () => {
   beforeEach(() => TestBed.configureTestingModule({
