+++ src/app/modules/games/games.module.ts
@@ -3,7 +3,7 @@ import { CommonModule } from '@angular/common';
 
 import { GamesRoutingModule } from './games.routing';
 
-import { SharedModule } from '@shared/shared-module.module';
+import { SharedModule } from '@shared/shared.module';
 
 import { GamesComponent } from './games.component';
 import { GameOptionsComponent, MainComponent } from './components';
