+++ src/app/modules/menu/menu.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { RouterModule } from '@angular/router';
-import { SharedModule } from '@shared/shared-module.module';
+import { SharedModule } from '@shared/shared.module';
 
 import { MenuComponent } from './menu.component';
 import { MenuRoutingModule } from './menu.routing';
